codebase_genius/requirements.txt
jaclang>=0.5.0
GitPython>=3.1.41
tree_sitter>=0.20.4
PyGithub>=2.2.0
python-dotenv>=1.0.1
rich>=13.7.0
graphviz>=0.20.3
networkx>=3.2.1
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
google-generativeai>=0.8.0
